Add parallel sieve

+++ C++/SieveOfEratosthenes/sieve.h
@@ -2,10 +2,10 @@
 #include <vector>
 
 namespace sieve {
-//Классический алгоритм
+// Классический алгоритм
 std::vector<unsigned> SieveOfEratosthenes(unsigned n);
 
-//Неограниченное решето Эратосфена
+// Неограниченное решето Эратосфена
 class UnlimitedSieve {
   std::vector<bool> primes;
   unsigned index;
@@ -18,23 +18,22 @@ public:
   unsigned nextPrime();
 };
 
-//Алгоритм с пропусками кратных
+// Алгоритм с пропусками кратных
 std::vector<unsigned> SieveOfEratosthenes_skip(unsigned n);
 
-//Алгоритм с кольцевой факторизацией
+// Алгоритм с кольцевой факторизацией
 std::vector<unsigned> SieveOfEratosthenes_wheel_factorization(unsigned n);
 
-//Сегментированное решето
+// Сегментированное решето
 std::vector<unsigned long long>
 SieveOfEratosthenes_segmented(unsigned long long n);
 
-//Алгоритм с линейным временем выполнения
+// Алгоритм с линейным временем выполнения
 std::vector<unsigned> SieveOfEratosthenes_linear(unsigned n);
 
-//Решето с распараллеливанием
-std::vector<unsigned long long>
-SieveOfEratosthenes_parallel(unsigned long long n);
+// Решето с распараллеливанием
+std::vector<unsigned> SieveOfEratosthenes_parallel(unsigned n);
 
-//Решето Аткина
+// Решето Аткина
 std::vector<unsigned long long> SieveOfAtkhin(unsigned long long n);
 } // namespace sieve
